2.0.0 : Chiffrement finis !

+++ src/AugmenterGraine.java
@@ -1,9 +1,9 @@
 public class AugmenterGraine {
-    private int[] graine;
+    private byte[] graine;
     private final int[] filtres;
     private int longueurGraine;
 
-    public AugmenterGraine(int[] graine, int[] filtres, int longueurGraine){
+    public AugmenterGraine(byte[] graine, int[] filtres, int longueurGraine){
         this.graine = graine;
         this.filtres = filtres;
         this.longueurGraine = longueurGraine;
@@ -11,11 +11,11 @@ public class AugmenterGraine {
 
     //Augmente la graine de 1 bit
     public void augmenterGraine(){
-        int bitARajouterALaFin = 0;
+        byte bitARajouterALaFin = 0;
 
         //Pour tous les filtres on fait un xor avec les bit correspondant à la taille de la graine moins le filtre
         for (int filtre : filtres) {
-            bitARajouterALaFin = bitARajouterALaFin ^ graine[(longueurGraine) - filtre];
+            bitARajouterALaFin = (byte) (bitARajouterALaFin ^ graine[(longueurGraine) - filtre]);
         }
 
         //On rajoute le bit de la fin au tableau
@@ -29,11 +29,11 @@ public class AugmenterGraine {
         return longueurGraine;
     }
 
-    public int getUnBitDeLaGraine(int numBit){
+    public byte getUnBitDeLaGraine(int numBit){
         return graine[numBit];
     }
 
-    public int[] getGraine() {
+    public byte[] getGraine() {
         return graine;
     }
 }
